+++ dart/geodata/lib/src/core/data/item_query.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2020-2023 Navibyte (https://navibyte.com). All rights reserved.
+// Copyright (c) 2020-2024 Navibyte (https://navibyte.com). All rights reserved.
 // Use of this source code is governed by a “BSD-3-Clause”-style license that is
 // specified in the LICENSE file.
 //
@@ -21,5 +21,17 @@ class ItemQuery extends GeospatialQuery {
   final Object id;
 
   @override
-  List<Object?> get props => [id, crs, parameters];
+  String toString() {
+    return '$id;$crs;$parameters';
+  }
+
+  @override
+  bool operator ==(Object other) =>
+      other is ItemQuery &&
+      id == other.id &&
+      crs == other.crs &&
+      parameters == other.parameters;
+
+  @override
+  int get hashCode => Object.hash(id, crs, parameters);
 }