+++ dart/geodata/lib/src/core/features/feature_item.dart
@@ -1,10 +1,9 @@
-// Copyright (c) 2020-2023 Navibyte (https://navibyte.com). All rights reserved.
+// Copyright (c) 2020-2024 Navibyte (https://navibyte.com). All rights reserved.
 // Use of this source code is governed by a “BSD-3-Clause”-style license that is
 // specified in the LICENSE file.
 //
 // Docs: https://github.com/navibyte/geospatial
 
-import 'package:equatable/equatable.dart';
 import 'package:geobase/vector_data.dart';
 import 'package:meta/meta.dart';
 
@@ -12,7 +11,7 @@ import '/src/common/meta/meta_aware.dart';
 
 /// A result from a feature source containing [feature] and [meta] data.
 @immutable
-class FeatureItem with MetaAware, EquatableMixin {
+class FeatureItem with MetaAware {
   /// Create a feature item instance with [feature] and optional [meta].
   const FeatureItem(this.feature, {Map<String, dynamic>? meta})
       : meta = meta ?? const {};
@@ -24,5 +23,14 @@ class FeatureItem with MetaAware, EquatableMixin {
   final Map<String, dynamic> meta;
 
   @override
-  List<Object?> get props => [feature, meta];
+  String toString() {
+    return '$feature;$meta';
+  }
+
+  @override
+  bool operator ==(Object other) =>
+      other is FeatureItem && feature == other.feature && meta == other.meta;
+
+  @override
+  int get hashCode => Object.hash(feature, meta);
 }