+++ dart/geodata/test/geodata_core_test.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2020-2023 Navibyte (https://navibyte.com). All rights reserved.
+// Copyright (c) 2020-2024 Navibyte (https://navibyte.com). All rights reserved.
 // Use of this source code is governed by a “BSD-3-Clause”-style license that is
 // specified in the LICENSE file.
 //
@@ -7,8 +7,6 @@
 // ignore_for_file: avoid_types_on_closure_parameters
 // ignore_for_file: prefer_const_constructors,require_trailing_commas
 // ignore_for_file: prefer_const_literals_to_create_immutables
-
-import 'package:equatable/equatable.dart';
 
 import 'package:geobase/common.dart';
 import 'package:geobase/coordinates.dart';
@@ -19,9 +17,6 @@ import 'package:geodata/src/utils/resolve_api_call.dart';
 import 'package:test/test.dart';
 
 void main() {
-  // configure Equatable to apply toString() default impls
-  EquatableConfig.stringify = true;
-
   group('Geodata "core/data" tests', () {
     setUp(() {
       // NOP